third_party/WebKit/LayoutTests/fast/canvas-api/canvas-negative-size-expected.txt

Test that canvas elements can't have negative height and/or width.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkDefaultWidth() is 300
PASS checkDefaultHeight() is 150
FAIL trySettingWidth('abc') should be 300. Was 0.
PASS trySettingWidth('200') is 200
PASS trySettingWidth('300') is 300
FAIL trySettingWidth(NaN) should be 300. Was 0.
FAIL trySettingWidth(Infinity) should be 300. Was 0.
FAIL trySettingWidth(null) should be 300. Was 0.
PASS trySettingWidth(true) is 1
PASS trySettingWidth(false) is 0
FAIL trySettingHeight('abc') should be 150. Was 0.
PASS trySettingHeight('200') is 200
PASS trySettingHeight('150') is 150
FAIL trySettingHeight(NaN) should be 150. Was 0.
FAIL trySettingHeight(Infinity) should be 150. Was 0.
FAIL trySettingHeight(null) should be 150. Was 0.
PASS trySettingHeight(true) is 1
PASS trySettingHeight(false) is 0
FAIL trySettingWidth('foo') should be 300. Was 0.
PASS trySettingWidthAttribute('foo') is 300
PASS tryCreatingCanvasWithWidth('foo') is 300
PASS trySettingWidth(-1) is 300
PASS trySettingWidthAttribute(-1) is 300
PASS tryCreatingCanvasWithWidth(-1) is 300
PASS trySettingWidth(0) is 0
PASS trySettingWidthAttribute(0) is 0
PASS tryCreatingCanvasWithWidth(0) is 0
PASS trySettingWidth(1) is 1
PASS trySettingWidthAttribute(1) is 1
PASS tryCreatingCanvasWithWidth(1) is 1
PASS trySettingWidth('+7') is 7
PASS trySettingWidthAttribute('+7') is 7
PASS tryCreatingCanvasWithWidth('+7') is 7
PASS trySettingWidth('-7') is 300
PASS trySettingWidthAttribute('-7') is 300
PASS tryCreatingCanvasWithWidth('-7') is 300
PASS trySettingWidth('123') is 123
PASS trySettingWidthAttribute('123') is 123
PASS tryCreatingCanvasWithWidth('123') is 123
FAIL trySettingHeight('foo') should be 150. Was 0.
PASS trySettingHeightAttribute('foo') is 150
PASS tryCreatingCanvasWithHeight('foo') is 150
PASS trySettingHeight(-1) is 150
PASS trySettingHeightAttribute(-1) is 150
PASS tryCreatingCanvasWithHeight(-1) is 150
PASS trySettingHeight(0) is 0
PASS trySettingHeightAttribute(0) is 0
PASS tryCreatingCanvasWithHeight(0) is 0
PASS trySettingHeight(1) is 1
PASS trySettingHeightAttribute(1) is 1
PASS tryCreatingCanvasWithHeight(1) is 1
PASS trySettingHeight('+7') is 7
PASS trySettingHeightAttribute('+7') is 7
PASS tryCreatingCanvasWithHeight('+7') is 7
PASS trySettingHeight('-7') is 150
PASS trySettingHeightAttribute('-7') is 150
PASS tryCreatingCanvasWithHeight('-7') is 150
PASS trySettingHeight('123') is 123
PASS trySettingHeightAttribute('123') is 123
PASS tryCreatingCanvasWithHeight('123') is 123
PASS successfullyParsed is true

TEST COMPLETE
